scx_layered: Implement dsp_local_scan
scx_layered uses per-LLC DSQs, which means that there are no L1/2 cache
locality considerations once a task is put into a DSQ. While the shared DSQ
is beneficial for work conservation and latency consistency, completely
ignoring L1/2 locality likely hurts for some workloads. Implement
dsp_local_scan which scans the DSQ for L1/2 local tasks and dispatches that
first. The scan depth can be set as a percentage of the number of CPUs in an
LLC and limited with a hard limit.