scx_rustland_core: Forbid mmap() syscall
The user-space schedulers should never perform blocking memory
allocations, otherwise the entire scheduling pipeline may get stuck.

To prevent this from happening, scx_rustland_core implements a
GlobalAlloc with a custom memory allocator that operates on a
pre-allocated locked memory arena and all the memory of the process is
automatically locked.

However, external libraries/crates can still execute mmap() syscalls
directly (i.e., libc), potentially stalling the scheduler, for example:

  R scx_rustland[159] -5016ms
      scx_state/flags=3/0x1 dsq_flags=0x0 ops_state/qseq=2/1
      sticky/holding_cpu=-1/-1 dsq_id=(n/a)
      dsq_vtime=0 slice=0 weight=100
      cpus=ff

    asm_sysvec_apic_timer_interrupt+0x1a/0x20
    mmap_region+0x65/0x140
    do_mmap+0x47d/0x620
    vm_mmap_pgoff+0xbc/0x1c0
    do_syscall_64+0xbb/0x1e0
    entry_SYSCALL_64_after_hwframe+0x77/0x7f

To catch these calls introduce a seccomp filter that returns EPERM when
the mmap() syscall is invoked.

This doesn't solve the problem, but it allows to catch the code that
invokes mmap() and we can exit early without having to wait for the
watchdog timeout.

Signed-off-by: Andrea Righi <arighi@nvidia.com>